+++ Firestore/CHANGELOG.md
@@ -1,5 +1,9 @@
 # 10.17.0
 - [feature] Add support for sum and average aggregate queries.
+- [feature] The `FirebaseFirestore` module now contains Firebase Firestore's
+  Swift-only APIs that were previously only available via the
+  `FirebaseFirestoreSwift` extension SDK. See the
+  `FirebaseFirestoreSwift` release note from this release for more details.
 
 # 10.16.0
 - [fixed] Fixed an issue where Firestore's binary SwiftPM distribution would